Code refactoring to create datasetDao
