+++ ambry-account/src/integration-test/java/com/github/ambry/account/MySqlAccountServiceIntegrationTest.java
@@ -16,6 +16,7 @@ package com.github.ambry.account;
 import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.account.AccountUtils.AccountUpdateInfo;
 import com.github.ambry.account.mysql.AccountDao;
+import com.github.ambry.account.mysql.DatasetDao;
 import com.github.ambry.account.mysql.MySqlAccountStore;
 import com.github.ambry.account.mysql.MySqlAccountStoreFactory;
 import com.github.ambry.config.ClusterMapConfig;
@@ -1338,7 +1339,7 @@ public class MySqlAccountServiceIntegrationTest {
     Statement statement = dbConnection.createStatement();
     statement.executeUpdate("DELETE FROM " + AccountDao.ACCOUNT_TABLE);
     statement.executeUpdate("DELETE FROM " + AccountDao.CONTAINER_TABLE);
-    statement.executeUpdate("DELETE FROM " + AccountDao.DATASET_TABLE);
-    statement.executeUpdate("DELETE FROM " + AccountDao.DATASET_VERSION_TABLE);
+    statement.executeUpdate("DELETE FROM " + DatasetDao.DATASET_TABLE);
+    statement.executeUpdate("DELETE FROM " + DatasetDao.DATASET_VERSION_TABLE);
   }
 }
